backend/chaingate_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
pyodbc==4.0.39
python-dotenv==1.0.0
python-jose==3.3.0
qrcode==7.4.2
Pillow==10.0.1
requests==2.31.0
cryptography==41.0.4
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
